+++ src/ByteDev.Crypto/EncodingTypeConverter.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace ByteDev.Crypto
+{
+    internal static class EncodingTypeConverter
+    {
+        public static Encoding.EncodingType ToEncodingLibType(EncodingType encodingType)
+        {
+            switch (encodingType)
+            {
+                case EncodingType.Base64:
+                    return Encoding.EncodingType.Base64;
+                case EncodingType.Hex:
+                    return Encoding.EncodingType.Hex;
+                default:
+                    throw new InvalidOperationException($"Unhandled Crypto encoding type: {encodingType}.");
+            }
+        }
+    }
+}